Added functionality for menu with ordering and price
Used addToOrder function to add selected menu items to the order list and update total price. Started by selecting the #order-items
element to list ordered items and the #order-total element to display the total price.
Then created a new <li> element for the ordered item and set its text to itemName and append
it to the #order-items list. Then fetched the item's price from the prices object and used Number() to turn the string of orderTotal.textContent in to a number.
The currentTotal from the #order-total element adds the items price to this total
and updates to the new total which is formatted to two decimal places to account for cents using currentTotal.toFixed(2) and
finally the initMenuSystem(menu) is called to start the menu system and display the menu items.